+++ include/rasterizor/Clipper.hpp
@@ -14,7 +14,7 @@ namespace Guarneri
 	{
 	public:
 		static std::vector<Triangle> cvv_clipping(const float& near_plane, const tinymath::Frustum& cvv, const Vertex& c1, const Vertex& c2, const Vertex& c3);
-		static void screen_clipping(Vertex& lhs, Vertex& rhs, const int& width);
+		static void screen_clipping(Vertex& lhs, Vertex& rhs, const size_t& width);
 		static bool inside_cvv(const tinymath::vec4f& c1, const tinymath::vec4f& c2, const tinymath::vec4f& c3);
 		static bool inside_cvv(const tinymath::vec4f& v);
 		static bool backface_culling_ndc(const tinymath::vec3f& c1, const tinymath::vec3f& c2, const tinymath::vec3f& c3);
