+++ include/rasterizor/GraphicsDevice.hpp
@@ -52,27 +52,27 @@ namespace Guarneri
 		std::queue<GraphicsCommand*> commands;
 		// framebuffer tiles
 		FrameTile* tiles;
-		int width;
-		int height;
-		uint32_t row_tile_count;
-		uint32_t col_tile_count;
-		uint32_t tile_length;
+		size_t width;
+		size_t height;
+		size_t row_tile_count;
+		size_t col_tile_count;
+		size_t tile_length;
 		uint8_t msaa_subsample_count;
 		uint8_t subsamples_per_axis;
 
 	public:
 		GraphicsDevice();
 		~GraphicsDevice();
-		void resize(uint32_t w, uint32_t h);
-		void initialize(uint32_t w, uint32_t h);
+		void resize(size_t w, size_t h);
+		void initialize(size_t w, size_t h);
 		void submit_draw_command(Shader* shader, const Vertex& v1, const Vertex& v2, const Vertex& v3, const tinymath::mat4x4& m, const tinymath::mat4x4& v, const tinymath::mat4x4& p);
 		void fence_draw_commands();
 		void present();
 		void clear_buffer(const BufferFlag& flag);
 		void set_subsample_count(const uint8_t& multiplier);
 		tinymath::color_rgba* get_framebuffer() { size_t size;  return framebuffer->get_ptr(size); };
-		int get_width() { return width; }
-		int get_height() { return height; }
+		size_t get_width() { return width; }
+		size_t get_height() { return height; }
 
 	public:
 		void draw_segment(const tinymath::vec3f& start, const tinymath::vec3f& end, const tinymath::Color& col, const tinymath::mat4x4& v, const tinymath::mat4x4& p, const tinymath::vec2f& screen_translation);
@@ -102,12 +102,12 @@ namespace Guarneri
 		void scanblock(const Triangle& tri, const Shader& shader);
 		void scanline(const Triangle& tri, const Shader& shader);
 		v2f process_vertex(const Shader& shader, const Vertex& vert) const;
-		void process_subsamples(RawBuffer<tinymath::color_rgba>* fbuf, RawBuffer<float>* zbuf, RawBuffer<uint8_t>* stencilbuf, const uint32_t& row, const uint32_t& col, const Shader& shader, const Vertex& v0, const Vertex& v1, const Vertex& v2, const float& area);
-		void process_fragment(RawBuffer<tinymath::color_rgba>* fbuf, RawBuffer<float>* zbuf, RawBuffer<uint8_t>* stencilbuf, const Vertex& v, const uint32_t& row, const uint32_t& col, const Shader& shader);
+		void process_subsamples(RawBuffer<tinymath::color_rgba>* fbuf, RawBuffer<float>* zbuf, RawBuffer<uint8_t>* stencilbuf, const size_t& row, const size_t& col, const Shader& shader, const Vertex& v0, const Vertex& v1, const Vertex& v2, const float& area);
+		void process_fragment(RawBuffer<tinymath::color_rgba>* fbuf, RawBuffer<float>* zbuf, RawBuffer<uint8_t>* stencilbuf, const Vertex& v, const size_t& row, const size_t& col, const Shader& shader);
 		bool validate_fragment(const PerSampleOperation& op_pass) const;
-		bool perform_stencil_test(RawBuffer<uint8_t>* stencilbuf, const uint8_t& ref_val, const uint8_t& read_mask, const CompareFunc& func, const uint32_t& row, const uint32_t& col) const;
-		void update_stencil_buffer(RawBuffer<uint8_t>* stencilbuf, const uint32_t& row, const uint32_t& col, const PerSampleOperation& op_pass, const StencilOp& stencil_pass_op, const StencilOp& stencil_fail_op, const StencilOp& stencil_zfail_op, const uint8_t& ref_val) const;
-		bool perform_depth_test(RawBuffer<float>* zbuf, const CompareFunc& func, const uint32_t& row, const uint32_t& col, const float& z) const;
+		bool perform_stencil_test(RawBuffer<uint8_t>* stencilbuf, const uint8_t& ref_val, const uint8_t& read_mask, const CompareFunc& func, const size_t& row, const size_t& col) const;
+		void update_stencil_buffer(RawBuffer<uint8_t>* stencilbuf, const size_t& row, const size_t& col, const PerSampleOperation& op_pass, const StencilOp& stencil_pass_op, const StencilOp& stencil_fail_op, const StencilOp& stencil_zfail_op, const uint8_t& ref_val) const;
+		bool perform_depth_test(RawBuffer<float>* zbuf, const CompareFunc& func, const size_t& row, const size_t& col, const float& z) const;
 		tinymath::Color blend(const tinymath::Color& src_color, const tinymath::Color& dst_color, const BlendFactor& src_factor, const BlendFactor& dst_factor, const BlendOp& op);
 		float linearize_depth(const float& depth, const float& near, const float& far) const;
 		float linearize_01depth(const float& depth, const float& near, const float& far) const;
